Move helpers to external file
